+++ src/Editor/Configuration/EditorConfiguration.cs
@@ -52,6 +52,8 @@ public class EditorConfiguration
 
     [JsonProperty("CurrentModelType")] public ModelType CurrentModelType = ModelType.Cube;
 
+    [JsonProperty("ModelScale")] public float ModelScale = 1f;
+
     public enum LightingPreset
     {
         SingleWhiteLight = 0,
@@ -63,6 +65,7 @@ public class EditorConfiguration
     [JsonProperty("CameraSettings")] public CameraSettings CameraSettings = new();
 
     [JsonProperty("OutputDirectoryPath")] public string OutputDirectoryPath = "";
+
 
     public void AddRecentFile(string filePath) =>
         CollectionHelpers.AddEntryToHistory(RecentFiles, filePath, MaxRecentFiles);