+++ src/Editor/Resources.cs
@@ -6,6 +6,7 @@ internal static class Resources
     public const string ResourceSkyBoxesFolderPath = $"{ResourceUiPath}/skybox";
     public const string ResourceToolBoarFolderPath = $"{ResourceUiPath}/toolbar";
 
+    public const string ResourcesPath = "resources";
     public const string ResourceModelsPath = "resources/models";
     public const string ResourceShaderFolderPath = "resources/shaders";
     private const string ResourceImageFolderPath = "resources/images";