+++ src/Editor/Windows/CodeWindow.cs
@@ -37,10 +37,10 @@ namespace Editor.Windows
                     isValid |= code.IsValid;
                 }
 
-                ImGui.BeginDisabled(needsRebuild == false);
+                //ImGui.BeginDisabled(needsRebuild == false);
                 if (ImGui.Button("Build"))
                     BuildPressed?.Invoke();
-                ImGui.EndDisabled();
+               // ImGui.EndDisabled();
 
 
                 ImGui.SameLine();