+++ src/Library/CodeVariable/CodeVariableTexture.cs
@@ -1,4 +1,5 @@
 ﻿using Newtonsoft.Json;
+using Raylib_cs;
 using System.ComponentModel.DataAnnotations;
 
 namespace Library.CodeVariable;
@@ -7,5 +8,5 @@ namespace Library.CodeVariable;
 public class CodeVariableTexture : CodeVariableBase
 {
     [Required][JsonProperty("Value")] public string Value { get; set; } = "";
-
+    [Required][JsonProperty("MaterialMapIndex")] public MaterialMapIndex? MaterialMapIndex { get; set; }
 }