+++ CHANGELOG.md
@@ -1,3 +1,10 @@
 # Changelog
 
 ## 0.1 in progress
+
+### Features
+
+### Known bugs
+
+* Switching from 4 to 1 light doesn't clear extra lights in shader
+